refactor: modularize landing page into separate components and add framer-motion dependency
